+++ clicommand/agent_start.go
@@ -85,6 +85,7 @@ type AgentStartConfig struct {
 	HealthCheckAddr            string   `cli:"health-check-addr"`
 	MetricsDatadog             bool     `cli:"metrics-datadog"`
 	MetricsDatadogHost         string   `cli:"metrics-datadog-host"`
+	TracingBackend             string   `cli:"tracing-backend"`
 	Spawn                      int      `cli:"spawn"`
 	LogFormat                  string   `cli:"log-format"`
 	CancelSignal               string   `cli:"cancel-signal"`
@@ -150,6 +151,14 @@ func DefaultConfigFilePaths() (paths []string) {
 	}
 
 	return
+}
+
+// validTracingBackends is a list of valid backends for tracing. This is sanity
+// checked during agent startup so that bootstrapped jobs don't silently have
+// no tracing if an invalid value is given.
+var validTracingBackends = map[string]struct{}{
+	"":        struct{}{},
+	"datadog": struct{}{},
 }
 
 var AgentStartCommand = cli.Command{
@@ -402,6 +411,12 @@ var AgentStartCommand = cli.Command{
 			EnvVars: []string{"BUILDKITE_REDACTED_VARS"},
 			Value:   cli.NewStringSlice("*_PASSWORD", "*_SECRET", "*_TOKEN"),
 		},
+		&cli.StringFlag{
+			Name:    "tracing-backend",
+			Usage:   "The name of the tracing backend to use.",
+			EnvVars: []string{"BUILDKITE_TRACING_BACKEND"},
+			Value:   "",
+		},
 
 		// API Flags
 		AgentRegisterTokenFlag,
@@ -567,6 +582,11 @@ var AgentStartCommand = cli.Command{
 			DatadogHost: cfg.MetricsDatadogHost,
 		})
 
+		// Sanity check supported tracing backends
+		if _, has := validTracingBackends[cfg.TracingBackend]; !has {
+			l.Fatal("The given tracing backend is not supported: %s", cfg.TracingBackend)
+		}
+
 		// AgentConfiguration is the runtime configuration for an agent
 		agentConf := agent.AgentConfiguration{
 			BootstrapScript:            cfg.BootstrapScript,
@@ -593,6 +613,7 @@ var AgentStartCommand = cli.Command{
 			Shell:                      cfg.Shell,
 			RedactedVars:               cfg.RedactedVars,
 			AcquireJob:                 cfg.AcquireJob,
+			TracingBackend:             cfg.TracingBackend,
 		}
 
 		if loader.File != nil {