+++ clicommand/bootstrap.go
@@ -54,6 +54,7 @@ type BootstrapConfig struct {
 	GitSubmodules                bool     `cli:"git-submodules"`
 	SSHKeyscan                   bool     `cli:"ssh-keyscan"`
 	AgentName                    string   `cli:"agent" validate:"required"`
+	Queue                        string   `cli:"queue"`
 	OrganizationSlug             string   `cli:"organization" validate:"required"`
 	PipelineSlug                 string   `cli:"pipeline" validate:"required"`
 	PipelineProvider             string   `cli:"pipeline-provider" validate:"required"`
@@ -81,6 +82,7 @@ type BootstrapConfig struct {
 	Phases                       []string `cli:"phases" normalize:"list"`
 	Profile                      string   `cli:"profile"`
 	RedactedVars                 []string `cli:"redacted-vars" normalize:"list"`
+	TracingBackend               string   `cli:"tracing-backend"`
 }
 
 var BootstrapCommand = cli.Command{
@@ -147,6 +149,12 @@ var BootstrapCommand = cli.Command{
 			Value:   "",
 			Usage:   "The name of the agent running the job",
 			EnvVars: []string{"BUILDKITE_AGENT_NAME"},
+		},
+		&cli.StringFlag{
+			Name:    "queue",
+			Value:   "",
+			Usage:   "The name of the queue the agent belongs to, if tagged",
+			EnvVars: []string{"BUILDKITE_AGENT_META_DATA_QUEUE"},
 		},
 		&cli.StringFlag{
 			Name:    "organization",
@@ -301,6 +309,12 @@ var BootstrapCommand = cli.Command{
 			Usage:   "Pattern of environment variable names containing sensitive values",
 			EnvVars: []string{"BUILDKITE_REDACTED_VARS"},
 		},
+		&cli.StringFlag{
+			Name:    "tracing-backend",
+			Usage:   "The name of the tracing backend to use.",
+			EnvVars: []string{"BUILDKITE_TRACING_BACKEND"},
+			Value:   "",
+		},
 		DebugFlag,
 		ExperimentsFlag,
 		ProfileFlag,
@@ -363,6 +377,7 @@ var BootstrapCommand = cli.Command{
 			GitCloneMirrorFlags:          cfg.GitCloneMirrorFlags,
 			GitCleanFlags:                cfg.GitCleanFlags,
 			AgentName:                    cfg.AgentName,
+			Queue:                        cfg.Queue,
 			PipelineProvider:             cfg.PipelineProvider,
 			PipelineSlug:                 cfg.PipelineSlug,
 			OrganizationSlug:             cfg.OrganizationSlug,
@@ -385,6 +400,7 @@ var BootstrapCommand = cli.Command{
 			Shell:                        cfg.Shell,
 			Phases:                       cfg.Phases,
 			RedactedVars:                 cfg.RedactedVars,
+			TracingBackend:               cfg.TracingBackend,
 		})
 
 		ctx, cancel := context.WithCancel(context.Background())